requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
PyGithub==2.1.1
slack-sdk==3.26.1
pydantic==2.5.2
typer==0.9.0
PyYAML==6.0.1
rich==13.7.0